refactor: 수익 정책 계산을 Payment 도메인으로 처리하도록 수정
