+++ kotlin-playground/src/main/kotlin/org/gonza/kotlinplayground/service/FinancialPolicy.kt
@@ -1,8 +1,10 @@
 package org.gonza.kotlinplayground.service
 
+import org.gonza.kotlinplayground.domain.payment.Payment
+
 interface FinancialPolicy {
     fun calculateReturnOnInvestment(
-        totalInvestment: Double,
-        profitAmount: Double,
+        totalInvestment: Payment,
+        profitAmount: Payment,
     ): Double
 }
