+++ kotlin-playground/src/test/kotlin/org/gonza/kotlinplayground/service/LottoFinancialPolicyTest.kt
@@ -1,5 +1,6 @@
 package org.gonza.kotlinplayground.service
 
+import org.gonza.kotlinplayground.domain.payment.Payment
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.Test
 
@@ -7,16 +8,16 @@ class LottoFinancialPolicyTest {
     @Test
     fun `계산된 수익률은 총 투자금액 대비 총 수익금의 백분율이어야 한다`() {
         val lottoFinancialPolicy = LottoFinancialPolicy()
-        val totalInvestment = 14000.0
-        val profitAmount = 5000.0
+        val totalInvestment = Payment(14000)
+        val profitAmount = Payment(5000)
 
-        val result = lottoFinancialPolicy.calculateReturnOnInvestment(
-            totalInvestment = totalInvestment,
-            profitAmount = profitAmount
-        )
+        val result =
+            lottoFinancialPolicy.calculateReturnOnInvestment(
+                totalInvestment = totalInvestment,
+                profitAmount = profitAmount,
+            )
 
         val expect = 0.3571
         assertEquals(result, expect, 0.001)
     }
 }
-
